refactor reducer
